# docs/openapi/b1.yaml
openapi: 3.1.0
info:
  title: LapanganGo Phase B2 API
  version: 2.0.0
  description: Kontrak kumulatif B1 dan B2 untuk runtime lokal sandbox LapanganGo.
...
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /notifications:
    get:
      tags: [Auth]
      operationId: listNotifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Notification feed for the active customer account.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AccountNotification" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{notificationId}/read:
    patch:
      tags: [Auth]
      operationId: markNotificationRead
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/PublicId" }
      responses:
        "204": { description: Notification marked as read. }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /business/pricing-rules:
    post:
      tags: [Business]
      operationId: createPriceRule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePriceRuleInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Price rule created after overlap validation.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { $ref: "#/components/schemas/PublicId" }
...
  /business/pricing-preview:
    post:
      tags: [Business]
      operationId: previewPricing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PricingPreviewInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "200":
          description: Resolved price and selected rule for every sample.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PricingPreviewResult" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /bookings/{bookingId}/cancel:
    post:
      tags: [Bookings]
      operationId: cancelCustomerBooking
      parameters:
        - {
            name: bookingId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/BookingReference" },
          }
        - {
            name: Idempotency-Key,
            in: header,
            required: true,
            schema: { type: string },
          }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [reason],
                    properties: { reason: { type: string } },
                  },
              },
          }
      responses:
        "200":
          {
            description: Cancellation decision,
            content:
              {
                application/json: { schema: { $ref: "#/components/schemas/B2Object" } },
              },
          }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
...
  /business/refunds:
    get:
      tags: [Finance]
      operationId: listBusinessRefunds
      parameters:
        - {
            name: tenantId,
            in: query,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      responses:
        "200":
          {
            description: Tenant refunds,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
  /business/bookings/{bookingId}/refunds:
    post:
      tags: [Finance]
      operationId: requestBusinessRefund
      parameters:
        - {
            name: bookingId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/BookingReference" },
          }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201":
          {
            description: Refund request created,
            content:
              {
                application/json: { schema: { $ref: "#/components/schemas/B2Object" } },
              },
          }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
...
  /admin/refunds/{refundId}/decision:
    post:
      tags: [Finance]
      operationId: decideAdminRefund
      parameters:
        - {
            name: refundId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Refund decision recorded. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /admin/refunds/{refundId}/retry:
    post:
      tags: [Finance]
      operationId: retryAdminRefund
      parameters:
        - {
            name: refundId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": { description: Failed refund queued again. }
...
  /admin/cancellation-policies:
    post:
      tags: [Finance]
      operationId: createCancellationPolicy
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201":
          {
            description: Policy template created,
            content:
              {
                application/json: { schema: { $ref: "#/components/schemas/B2Object" } },
              },
          }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /business/venues/{venueId}/cancellation-policy:
    put:
      tags: [Finance]
      operationId: assignVenueCancellationPolicy
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Venue policy assigned. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /notifications/preferences:
    get:
      tags: [Auth]
      operationId: listNotificationPreferences
      responses:
        "200":
          {
            description: User notification preferences,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
    put:
      tags: [Auth]
      operationId: updateNotificationPreference
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Preference updated. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
...
  /admin/notification-reminder-options:
    post:
      tags: [Auth]
      operationId: createNotificationReminderOption
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201":
          {
            description: Reminder option created,
            content:
              {
                application/json: { schema: { $ref: "#/components/schemas/B2Object" } },
              },
          }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /business/venues/{venueId}/reminders:
    put:
      tags: [Auth]
      operationId: updateVenueReminders
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Venue reminder selections replaced. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
...
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 160 }
...
    VenueId:
      name: venueId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PublicId" }
...
  responses:
    Unauthorized:
      description: Authentication is required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    PayloadTooLarge:
      description: JSON request body exceeds the global 1 MB limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
    B2Object:
      type: object
      additionalProperties: true
    B2List:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/B2Object" }
...
    PublicId:
      type: string
      pattern: "^[A-Za-z0-9_-]{22}$"
      description: Deterministically encrypted reference; internal numeric key is never exposed.
    BookingReference:
      type: string
      pattern: "^LG-[A-Za-z0-9_-]{16}$"
      description: Opaque random reference used in customer-facing booking URLs.
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
        requestId: { type: string }
...
    AccountNotification:
      type: object
      required: [id, kind, title, body, actionPath, read, createdAt]
      properties:
        id: { $ref: "#/components/schemas/PublicId" }
        kind: { type: string, enum: [booking, payment, verification, system] }
        title: { type: string }
        body: { type: string }
        actionPath: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    CreatePriceRuleInput:
      type: object
      required:
        [
          tenantId,
          venueId,
          courtId,
          kind,
          amount,
          dayOfWeek,
          specialDate,
          startsAtLocal,
          endsAtLocal,
        ]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        venueId: { $ref: "#/components/schemas/PublicId" }
        courtId: { type: [string, "null"] }
        kind: { type: string, enum: [BASE, WEEKDAY_WEEKEND, DAY_TIME, SPECIAL_DATE] }
        amount: { type: integer, minimum: 1 }
        dayOfWeek: { type: [integer, "null"], minimum: 0, maximum: 6 }
        specialDate: { type: [string, "null"], format: date }
        startsAtLocal: { type: [string, "null"] }
        endsAtLocal: { type: [string, "null"] }
...
    PricingPreviewInput:
      type: object
      required: [tenantId, venueId, courtId, samples]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        venueId: { $ref: "#/components/schemas/PublicId" }
        courtId: { $ref: "#/components/schemas/PublicId" }
        candidate: { $ref: "#/components/schemas/CreatePriceRuleInput" }
        samples:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: object
            required: [localDate, localTime]
            properties:
              localDate: { type: string, format: date }
              localTime:
                { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d$" }
    PricingPreviewResult:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required:
              [localDate, localTime, amount, selectedRuleId, selectedKind, scope]
            properties:
              localDate: { type: string, format: date }
              localTime: { type: string }
              amount: { type: integer }
              selectedRuleId:
                type: string
                description: Public rule ID, or `candidate` for an unsaved preview rule.
              selectedKind:
                { type: string, enum: [BASE, WEEKDAY_WEEKEND, DAY_TIME, SPECIAL_DATE] }
              scope: { type: string, enum: [COURT, VENUE] }
...